+++ src/main/java/br/com/brq/challenges/mercadinho/entrypoint/controller/DepartamentoController.java
@@ -46,7 +46,7 @@ public class DepartamentoController {
 
     @DeleteMapping("/{id_departamento}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removerDepartamento(@PathVariable("id_departamento") Long idDepartamento) {
+    public void removerDepartamento(@PathVariable("id_departamento") Integer idDepartamento) {
         departamentoUseCase.removerDepartamento(idDepartamento);
     }
 }
